requirements.txt
pytest
allure-pytest
pytest-xdist
pytest-schema